Move suggestion_validator logic into own module
