+++ core/domain/prod_validators.py
@@ -21,7 +21,6 @@ from __future__ import unicode_literals  # pylint: disable=import-only-modules
 
 import datetime
 import itertools
-import re
 
 from constants import constants
 from core.domain import base_model_validators
@@ -89,9 +88,6 @@ TARGET_TYPE_TO_TARGET_MODEL = {
     suggestion_models.TARGET_TYPE_TOPIC: (
         topic_models.TopicModel)
 }
-VALID_SCORE_CATEGORIES_FOR_TYPE_QUESTION = [
-    '%s\\.[A-Za-z0-9-_]{1,%s}' % (
-        suggestion_models.SCORE_TYPE_QUESTION, base_models.ID_LENGTH)]
 
 
 class ExplorationModelValidator(base_model_validators.BaseModelValidator):
@@ -980,233 +976,6 @@ class ContinuousComputationModelValidator(
     @classmethod
     def _get_custom_validation_functions(cls):
         return [cls._validate_time_fields]
-
-
-class GeneralSuggestionModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating GeneralSuggestionModels."""
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        # Valid id: same as thread id:
-        # [target_type].[target_id].[GENERATED_STRING].
-        regex_string = '^%s\\.%s\\.[A-Za-z0-9=+/]{1,}$' % (
-            item.target_type, item.target_id)
-        return regex_string
-
-    @classmethod
-    def _get_model_domain_object_instance(cls, item):
-        if item.target_type in TARGET_TYPE_TO_TARGET_MODEL:
-            return suggestion_services.get_suggestion_from_model(item)
-        else:
-            cls._add_error(
-                'target %s' % base_model_validators.ERROR_CATEGORY_TYPE_CHECK,
-                'Entity id %s: Target type %s is not allowed' % (
-                    item.id, item.target_type))
-            return None
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        field_name_to_external_model_references = [
-            base_model_validators.ExternalModelFetcherDetails(
-                'feedback_thread_ids',
-                feedback_models.GeneralFeedbackThreadModel, [item.id])
-        ]
-        if user_services.is_user_id_valid(item.author_id):
-            field_name_to_external_model_references.append(
-                base_model_validators.ExternalModelFetcherDetails(
-                    'author_ids',
-                    user_models.UserSettingsModel,
-                    [item.author_id]
-                )
-            )
-        if item.target_type in TARGET_TYPE_TO_TARGET_MODEL:
-            field_name_to_external_model_references.append(
-                base_model_validators.ExternalModelFetcherDetails(
-                    '%s_ids' % item.target_type,
-                    TARGET_TYPE_TO_TARGET_MODEL[item.target_type],
-                    [item.target_id]))
-        if item.final_reviewer_id and user_services.is_user_id_valid(
-                item.final_reviewer_id):
-
-            # Bot rejects suggestions when the suggestion's targeted entity gets
-            # removed from the topic. The bot doesn't have a UserSettingsModel
-            # for their user_id. Exclude external model validation for bot.
-            if item.final_reviewer_id != feconf.SUGGESTION_BOT_USER_ID:
-                field_name_to_external_model_references.append(
-                    base_model_validators.ExternalModelFetcherDetails(
-                        'reviewer_ids', user_models.UserSettingsModel,
-                        [item.final_reviewer_id]))
-        return field_name_to_external_model_references
-
-    @classmethod
-    def _validate_target_type(cls, item):
-        """Validate the target type is valid.
-
-        Args:
-            item: datastore_services.Model. GeneralSuggestionModel to validate.
-        """
-        if item.target_type not in TARGET_TYPE_TO_TARGET_MODEL:
-            cls._add_error(
-                'target %s' % base_model_validators.ERROR_CATEGORY_TYPE_CHECK,
-                'Entity id %s: Target type %s is not allowed' % (
-                    item.id, item.target_type))
-
-    @classmethod
-    def _validate_target_version_at_submission(
-            cls, item, field_name_to_external_model_references):
-        """Validate the target version at submission is less than or
-        equal to the version of the target model.
-
-        Args:
-            item: datastore_services.Model. GeneralSuggestionModel to validate.
-            field_name_to_external_model_references:
-                dict(str, (list(base_model_validators.ExternalModelReference))).
-                A dict keyed by field name. The field name represents
-                a unique identifier provided by the storage
-                model to which the external model is associated. Each value
-                contains a list of ExternalModelReference objects corresponding
-                to the field_name. For examples, all the external Exploration
-                Models corresponding to a storage model can be associated
-                with the field name 'exp_ids'. This dict is used for
-                validation of External Model properties linked to the
-                storage model.
-        """
-        if item.target_type not in TARGET_TYPE_TO_TARGET_MODEL:
-            cls._add_error(
-                'target %s' % base_model_validators.ERROR_CATEGORY_TYPE_CHECK,
-                'Entity id %s: Target type %s is not allowed' % (
-                    item.id, item.target_type))
-            return
-
-        target_model_references = (
-            field_name_to_external_model_references[
-                '%s_ids' % item.target_type])
-
-        for target_model_reference in target_model_references:
-            target_model = target_model_reference.model_instance
-            if target_model is None or target_model.deleted:
-                model_class = target_model_reference.model_class
-                model_id = target_model_reference.model_id
-                cls._add_error(
-                    '%s_ids %s' % (
-                        item.target_type,
-                        base_model_validators.ERROR_CATEGORY_FIELD_CHECK),
-                    'Entity id %s: based on field %s_ids having'
-                    ' value %s, expected model %s with id %s but it doesn\'t'
-                    ' exist' % (
-                        item.id, item.target_type,
-                        model_id, model_class.__name__, model_id))
-                continue
-            if item.target_version_at_submission > target_model.version:
-                cls._add_error(
-                    'target version at submission check',
-                    'Entity id %s: target version %s in entity is greater '
-                    'than the version %s of %s corresponding to '
-                    'id %s' % (
-                        item.id, item.target_version_at_submission,
-                        target_model.version, item.target_type, item.target_id))
-
-    @classmethod
-    def _validate_final_reviewer_id(cls, item):
-        """Validate that final reviewer id is None if suggestion is
-        under review.
-
-        Args:
-            item: datastore_services.Model. GeneralSuggestionModel to validate.
-        """
-        if item.final_reviewer_id is None and (
-                item.status != suggestion_models.STATUS_IN_REVIEW):
-            cls._add_error(
-                'final %s' % (
-                    base_model_validators.ERROR_CATEGORY_REVIEWER_CHECK),
-                'Entity id %s: Final reviewer id is empty but '
-                'suggestion is %s' % (item.id, item.status))
-
-        if item.final_reviewer_id:
-            if item.status == suggestion_models.STATUS_IN_REVIEW:
-                cls._add_error(
-                    'final %s' % (
-                        base_model_validators.ERROR_CATEGORY_REVIEWER_CHECK),
-                    'Entity id %s: Final reviewer id %s is not empty but '
-                    'suggestion is in review' % (
-                        item.id, item.final_reviewer_id))
-
-    @classmethod
-    def _validate_score_category(
-            cls, item, field_name_to_external_model_references):
-        """Validate that the score_category subtype for suggestions matches the
-        exploration category.
-
-        Args:
-            item: datastore_services.Model. GeneralSuggestionModel to validate.
-            field_name_to_external_model_references:
-                dict(str, (list(base_model_validators.ExternalModelReference))).
-                A dict keyed by field name. The field name represents
-                a unique identifier provided by the storage
-                model to which the external model is associated. Each value
-                contains a list of ExternalModelReference objects corresponding
-                to the field_name. For examples, all the external Exploration
-                Models corresponding to a storage model can be associated
-                with the field name 'exp_ids'. This dict is used for
-                validation of External Model properties linked to the
-                storage model.
-        """
-        if item.target_type not in TARGET_TYPE_TO_TARGET_MODEL:
-            cls._add_error(
-                'target %s' % base_model_validators.ERROR_CATEGORY_TYPE_CHECK,
-                'Entity id %s: Target type %s is not allowed' % (
-                    item.id, item.target_type))
-            return
-        score_category_type = (
-            item.score_category.split(
-                suggestion_models.SCORE_CATEGORY_DELIMITER)[0])
-
-        if item.target_type == suggestion_models.TARGET_TYPE_EXPLORATION:
-            target_model_references = (
-                field_name_to_external_model_references[
-                    '%s_ids' % item.target_type])
-
-            for target_model_reference in target_model_references:
-                target_model = target_model_reference.model_instance
-                if target_model is None or target_model.deleted:
-                    model_class = target_model_reference.model_class
-                    model_id = target_model_reference.model_id
-                    cls._add_error(
-                        '%s_ids %s' % (
-                            item.target_type,
-                            base_model_validators.ERROR_CATEGORY_FIELD_CHECK),
-                        'Entity id %s: based on field %s_ids having'
-                        ' value %s, expected model %s with id %s but it '
-                        'doesn\'t exist' % (
-                            item.id, item.target_type,
-                            model_id, model_class.__name__, model_id))
-
-                # Note: An exploration's category can be changed after the
-                # suggestion is submitted. Since this operation does not update
-                # the suggestion's category, we cannot assume that the
-                # exploration category matches the suggestion score category,
-                # and thus do not validate it here.
-
-        if score_category_type == suggestion_models.SCORE_TYPE_QUESTION:
-            score_category_regex = (
-                '^(%s)$' % ('|').join(VALID_SCORE_CATEGORIES_FOR_TYPE_QUESTION))
-            if not re.compile(score_category_regex).match(item.score_category):
-                cls._add_error(
-                    'score category check',
-                    'Entity id %s: Score category %s is invalid' % (
-                        item.id, item.score_category))
-
-    @classmethod
-    def _get_custom_validation_functions(cls):
-        return [
-            cls._validate_target_type,
-            cls._validate_final_reviewer_id]
-
-    @classmethod
-    def _get_external_instance_custom_validation_functions(cls):
-        return [
-            cls._validate_target_version_at_submission,
-            cls._validate_score_category]
 
 
 class GeneralVoiceoverApplicationModelValidator(
